feat(tests/dummy/app/templates/catalog): adds grid list.
